CLN: 포스팅 전체 조회 리팩토링 및 응답값 수정

+++ src/main/java/kr/hoppang/domain/boards/repository/dto/ConditionOfFindPosts.java
@@ -6,6 +6,6 @@ import lombok.Builder;
 @Builder
 public record ConditionOfFindPosts(
         List<Long> boardIds,
-        long limit,
+        int limit,
         long offset
 ) { }
